feat(aia): add implementations for APLIC peripheral registers

+++ library/riscv-aia/src/register/hviprio2.rs
@@ -10,7 +10,7 @@ impl Hviprio2 {
     /// Returns the priority byte at byte index `i` (0..7).
     #[inline]
     pub const fn prio_byte(self, i: usize) -> u8 {
-        let shift = (i as usize) * 8;
+        let shift = i * 8;
         ((self.bits >> shift) & 0xFF) as u8
     }
 }
